add bootrom to support downloading software from the uart port

+++ sw/tc_ext_access/main.c
@@ -1,7 +1,7 @@
 
 #include <stdint.h>
 
-extern volatile uint64_t fromhost;
+extern volatile uint32_t fromhost;
 
 int main()
 {
